search: expression/constraint aliases
The search handler can store and retrieve search aliases.
Keyword namedSearch handles these new atoms of the form
named:foo.
Review could start with the question whether the core of NamedSearch
was implemented in the right place.
The core consists of search handler methods GetNamed and SetNamed, for
lookup and definition of aliases.
I implemented them in the search handler because:
1. NamedSearch depends on search for lookup and evaluation
2. gives /camli/search/{set,get}named urls
This comes at a price because SetNamed needs a storageAndIndex
component.
Which gives rise to one more dependency for search handler in the
configs.

When the core has been reviewed you can continue with
NamedSearch in predicate.go

The changes beside handler.go and predicate.go are less interesting.
I needed to make a camtool subcommand for end to end testing. Going
through the client.Client seemed obvious.

Change-Id: I7960f83bad464eb1a971c07f33631744a5eea814